test(top-level-await): move out of staging module graphs hanging
