+++ test/language/module-code/top-level-await/module-graphs-grandparent-tla_FIXTURE.js
@@ -1,4 +1,4 @@
 // Copyright (C) 2024 the V8 project authors. All rights reserved.
 // This code is governed by the BSD license found in the LICENSE file.
 
-import "./parent-tla_FIXTURE.js";
+import "./module-graphs-parent-tla_FIXTURE.js";
