+++ test/language/module-code/top-level-await/tla_FIXTURE.js
